chore: update the example basic-app
