+++ examples/basic-app/src-tauri/gen/schemas/capabilities.json
@@ -1,0 +1,1 @@
+{"main":{"identifier":"main","description":"applied permissions for our demo application","local":true,"windows":["main"],"permissions":["sentry:default"]}}
